[Silver III] Title: 주유소, Time: 28 ms, Memory: 4364 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.cc
@@ -2,9 +2,9 @@
 #define MAX 100001
 using namespace std;
 
-int dist[MAX];
-int city[MAX];
-int city_minimun[MAX];
+long long dist[MAX];
+long long city[MAX];
+long long city_minimun[MAX];
 
 int main() {
   ios::sync_with_stdio(false);
